# docs/services/gateway/openapi.yaml
openapi: 3.1.0
info:
  title: CORTX Gateway
  version: 0.1.0
paths:
...
  /jobs/validate:
    post:
      tags:
      - Orchestrator
      summary: Create Validation Job
      description: Create and execute validation job.
      operationId: create_validation_job_jobs_validate_post
      parameters:
      - name: domain
        in: query
        required: true
        schema:
          type: string
          title: Domain
      - name: mode
        in: query
        required: false
        schema:
          type: string
          default: static
          title: Mode
      - name: input_ref
        in: query
        required: false
        schema:
          type: string
          title: Input Ref
      - name: x-tenant-id
        in: header
        required: false
        schema:
          type: string
          default: default
          title: X-Tenant-Id
      - name: x-request-id
        in: header
        required: false
        schema:
          type: string
          title: X-Request-Id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_create_validation_job_jobs_validate_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Validation Job Jobs Validate Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /jobs/{job_id}:
    get:
      tags:
      - Orchestrator
      summary: Get Job Status
      description: Get validation job status and results.
      operationId: get_job_status_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      - name: x-tenant-id
        in: header
        required: false
        schema:
          type: string
          default: default
          title: X-Tenant-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Job Status Jobs  Job Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /explain:
    post:
      tags:
      - Orchestrator
      summary: Explain Failure
      description: Get explanation for a validation failure.
      operationId: explain_failure_explain_post
      parameters:
      - name: domain
        in: query
        required: true
        schema:
          type: string
          title: Domain
      - name: failure_id
        in: query
        required: true
        schema:
          type: string
          title: Failure Id
      - name: include_policy_refs
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Policy Refs
      - name: include_remediation
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Remediation
      - name: x-tenant-id
        in: header
        required: false
        schema:
          type: string
          default: default
          title: X-Tenant-Id
      - name: x-request-id
        in: header
        required: false
        schema:
          type: string
          title: X-Request-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Failure Data
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Explain Failure Explain Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /failures/{failure_id}/decision:
    put:
      tags:
      - Orchestrator
      summary: Update Failure Decision
      description: Update user decision on a validation failure.
      operationId: update_failure_decision_failures__failure_id__decision_put
      parameters:
      - name: failure_id
        in: path
        required: true
        schema:
          type: string
          title: Failure Id
      - name: decision
        in: query
        required: true
        schema:
          type: string
          title: Decision
      - name: reason
        in: query
        required: false
        schema:
          type: string
          title: Reason
      - name: notes
        in: query
        required: false
        schema:
          type: string
          title: Notes
      - name: x-tenant-id
        in: header
        required: false
        schema:
          type: string
          default: default
          title: X-Tenant-Id
      - name: x-request-id
        in: header
        required: false
        schema:
          type: string
          title: X-Request-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Update Failure Decision Failures  Failure Id  Decision
                  Put
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /feedback/rag/{interaction_id}:
    post:
      tags:
      - Orchestrator
      summary: Submit Rag Feedback
      description: Submit feedback on RAG/AI explanations.
      operationId: submit_rag_feedback_feedback_rag__interaction_id__post
      parameters:
      - name: interaction_id
        in: path
        required: true
        schema:
          type: string
          title: Interaction Id
      - name: feedback
        in: query
        required: true
        schema:
          type: string
          title: Feedback
      - name: details
        in: query
        required: false
        schema:
          type: string
          title: Details
      - name: x-tenant-id
        in: header
        required: false
        schema:
          type: string
          default: default
          title: X-Tenant-Id
      - name: x-request-id
        in: header
        required: false
        schema:
          type: string
          title: X-Request-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Submit Rag Feedback Feedback Rag  Interaction Id  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Body_create_validation_job_jobs_validate_post:
      properties:
        input_data:
          additionalProperties: true
          type: object
          title: Input Data
        options:
          additionalProperties: true
          type: object
          title: Options
      type: object
      title: Body_create_validation_job_jobs_validate_post
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError